Buffs : Added feature to display buff sources
I was forever forgetting which class/spec gave each buff so I've added a useful OPTIONAL feature to Rawr. In Tools.. Options.. General Settings is a new option to enable or disable displaying of Buff Sources. When enabled the buff panel displays the name of the buff and the source, when disabled its exactly the same as before. Default is disabled.

To add sources there is a new optional string field in the Buff class called Source. Simply populate this to use the feature.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C32707