Tweaking the TankDK model slightly.  I had some more drastic changes, but pruned them back since I didn't have reasonable tests for them yet.  Not sure on the relative stat numbers and weights, but they look a bit closer what I would expect from my current gear and where I *think* I should be going in my next upgrades.
Still alot more to come as I migrate to using some of the updated constants that are in the Rawr.Base areas.  I just want some tests before making those kinds of drastic transitions.

Also - Adding a Readme.TankDK.txt for tracking the current state of the module with what I believe is working and not.  

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C34074