Additional Talents modeled in, but no where close to being done.  Really this calls for a major re-write to look closer to the ProtWarrior stuff.  Allowing the user to tune the type of encounter and adjust the levels and types of damage.  As well as setting up a shot rotation that works for the spec since each one has their own.
Added the option of adjusting the ammount of IncomingDamage the tank is tanking.  This is basically a way to tune Mitigation.  The previous method of Survival vs. Threat tuning wasn't suiting my needs. With this however, I'm seeing that DEF is very heavily weighted even after the crit cap.  So I need to look into that.  I would rather expect to see that DEF's value decrease, and other mitigation stats like Dodge and Parry over take it.  

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C34301