Rawr.Base: - Migrated NPC Boss Armor stats to an array, checkable against Boss Level - 80, updated models to reflect this change and populated Target Armor combo boxes with this as a list
Rogue, ProtPally, Retribution, DPSDK have each had an overhaul to use 3.2 Ratings Conversions. I still need to overhaul the caster classes and TankDK

NOTICE: Model Devs need to run their models and very no NULL exception errors. When adding the TargetLevel-80 reference I tried to catch all the -80's but there may be one or two still in the code that need to be added, otherwise it will try to check the 83rd point in a 4 point array.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C36151