Enhance : Implemented Paragon and removed GreatnessProc re-coding it for HighestStat
As Elemental was only model still using GreatnessProc I modified a couple of lines in Rawr.Elemental as well and removed the GreatnessProc re-coding it for highest stat. As GreatnessProc was then completely redundant I removed it from Stats and the Darkmoon Card Greatness. Thus eliminating the redundant code completely.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C36776