Applying Patch 10501:
Pertains to Issue 20990: High dpi support; wrapping character slots and cut off search boxes

Adds a ScaleTransform to the Left and Right column of Item Buttons. When Compact Paper Doll Mode is active User can Scale down these buttons by a percentage to let them be smaller and no longer wrap to next column. These wraps cause the Charts to become so squished as to be unusable.

Includes sample image comparing a 1024x728 space with and without Compact Paper Doll Mode @85%.

Also corrects the spacing for the Status Bar messaging, which was overlapping and cutting off the text.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C63324